Rename manual mod action module
